chore: remove proto2 special-casing from mappers for compute APIs
